project/requirements.txt
amqp==2.6.1
asgiref==3.4.1
billiard==3.5.0.5
celery==4.2.1
Django==2.1
djangorestframework==3.12.4
kombu==4.6.11
mysqlclient==2.0.3
pytz==2021.1
sqlparse==0.4.2
vine==1.3.0

certifi==2021.5.30
charset-normalizer==2.0.6
idna==3.2
requests==2.26.0
urllib3==1.26.7

